Code refactoring to support the Integration Tests created for F# requests.

+++ RepoDb.Core/RepoDb.Tests/RepoDb.IntegrationTests/Helper.cs
@@ -756,7 +756,7 @@ namespace RepoDb.IntegrationTests
                 ColumnDateTime = EpocDate.AddDays(1),
                 ColumnDateTime2 = DateTime.UtcNow,
                 ColumnDecimal = Convert.ToDecimal(1),
-                ColumnFloat = Convert.ToSingle(1),
+                ColumnFloat = Convert.ToDouble(1),
                 ColumnInt = 1,
                 ColumnNVarChar = $"NVARCHAR{1}"
             };
@@ -781,7 +781,7 @@ namespace RepoDb.IntegrationTests
                     ColumnDateTime = EpocDate.AddDays(index),
                     ColumnDateTime2 = DateTime.UtcNow,
                     ColumnDecimal = Convert.ToDecimal(index),
-                    ColumnFloat = Convert.ToSingle(index),
+                    ColumnFloat = Convert.ToDouble(index),
                     ColumnInt = index,
                     ColumnNVarChar = $"NVARCHAR{index}"
                 });
@@ -837,7 +837,7 @@ namespace RepoDb.IntegrationTests
                     ColumnDateTime = EpocDate.AddDays(index),
                     ColumnDateTime2 = DateTime.UtcNow,
                     ColumnDecimal = Convert.ToDecimal(index),
-                    ColumnFloat = Convert.ToSingle(index),
+                    ColumnFloat = Convert.ToDouble(index),
                     ColumnInt = index,
                     ColumnNVarChar = $"NVARCHAR{index}"
                 });
